+++ autorest/codegen/models/__init__.py
@@ -29,6 +29,7 @@ from .lro_paging_operation import LROPagingOperation
 
 __all__ = [
     "AzureKeyCredentialSchema",
+    "AnySchema",
     "BaseModel",
     "BaseSchema",
     "CodeModel",