+++ autorest/codegen/models/parameter.py
@@ -199,7 +199,7 @@ class Parameter(BaseModel):  # pylint: disable=too-many-instance-attributes
     @property
     def in_method_code(self) -> bool:
         return not (
-            isinstance(self.schema, ConstantSchema) and
+            self.constant and
             self.location == ParameterLocation.Other or
             self.rest_api_name == '$host'
         )