+++ test/vanilla/legacy/Expected/AcceptanceTests/Anything/anything/aio/__init__.py
@@ -6,14 +6,6 @@
 # Changes may cause incorrect behavior and will be lost if the code is regenerated.
 # --------------------------------------------------------------------------
 
-try:
-    from ._request_builders_py3 import build_get_request
-    from ._request_builders_py3 import build_put_request
-except (SyntaxError, ImportError):
-    from ._request_builders import build_get_request  # type: ignore
-    from ._request_builders import build_put_request  # type: ignore
+from ._anything_client import AnythingClient
 
-__all__ = [
-    "build_get_request",
-    "build_put_request",
-]
+__all__ = ["AnythingClient"]