+++ test/vanilla/low-level/Expected/AcceptanceTests/ObjectTypeLowLevel/objecttypelowlevel/rest/_request_builders_py3.py
@@ -62,7 +62,7 @@ def build_put_request(*, json: Any = None, content: Any = None, **kwargs: Any) -
         .. code-block:: python
 
             # JSON input template you can fill out and use as your `json` input.
-            json = "str (optional)"
+            json = "object (optional)"
     """
 
     content_type = kwargs.pop("content_type", None)  # type: Optional[str]